requirements.txt
asgiref==3.8.1
boto3==1.36.2
botocore==1.36.2
Django==5.1.4
django-storages==1.14.4
django-taggit==6.1.0
django-tinymce==4.1.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.1
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.8.2
wonderwords==2.2.0
